auth reducers and actions moved to proper directories
